clean Seni`s code add library class
