+++ Libirary_/src/main/java/com/example/libirary_/User.java
@@ -4,12 +4,14 @@ import userprofile.Order;
 
 import java.util.ArrayList;
 
+import static librarypackage.Library.users;
+
 public class User extends Person{
     private static int userCounter =0;
     private static User currentUser;
-   public static ArrayList<User> users = new ArrayList<User>();
 
-   public ArrayList<Order> orders = new ArrayList<Order>();
+  private ArrayList<Integer>userOrders = new ArrayList<>();//order id
+
 
    public User(int id,String userName, String email, String password){
 
@@ -30,6 +32,17 @@ public class User extends Person{
     public static void setCurrentUser(User user){
         currentUser = user;
     }
+    public void addOrder(int orderID){
+       this.userOrders.add(orderID);
+    }
+    public boolean findOrderID(int orderId){
 
+       for(Integer id : this.userOrders){
+               if(orderId==id)
+                   return true;
+       }
+
+       return false;
+    }
 
 }
