+++ Libirary_/src/main/java/librarypackage/Library.java
@@ -1,0 +1,11 @@
+package librarypackage;
+
+import com.example.libirary_.User;
+import userprofile.Order;
+
+import java.util.ArrayList;
+
+public  class Library {
+    public static ArrayList<User> users = new ArrayList<User>();
+    public static ArrayList<Order> orders = new ArrayList<Order>();
+}
